// src/services/api/generated/movie-service/model/createGenreRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Cinema Booking System API
 * API documentation for Cinema Booking System
 * OpenAPI spec version: 1.0.0
 */

export interface CreateGenreRequest {
  /**
     * @minLength 0
     * @maxLength 100
     */
  name: string;
  /**
     * @minLength 0
     * @maxLength 500
     */
  description?: string;
}