// src/services/api/generated/movie-service/model/genreResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Cinema Booking System API
 * API documentation for Cinema Booking System
 * OpenAPI spec version: 1.0.0
 */

export interface GenreResponse {
  id?: string;
  name?: string;
  description?: string;
  version?: number;
  createdAt?: string;
  updatedAt?: string;
}